+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - VectorTrait for providing information about math vector types (**@RiscadoA**).
 - Support for VectorTrait on the inspector (**@RiscadoA**).
 - Support for quaternions on the inspector (#1018, **@RiscadoA**).
+- Decomposition of 4x4 matrices into translation, scale and rotation on the inspector (#1179, **@RiscadoA**).
 
 ### Changed